+++ src/main/java/com/example/goldenticketnew/repository/TicketRepository.java
@@ -11,5 +11,5 @@ public interface TicketRepository extends JpaRepository<Ticket, Integer> {
     List<Ticket> findTicketsBySchedule_Id(Integer scheduleId);
     List<Ticket> findTicketsBySchedule_IdAndSeat_Id(Integer scheduleId,Integer seatId);
     @Query("SELECT t FROM Ticket t WHERE t.bill.id IN (SELECT b.id FROM Bill b WHERE b.user.id=:userId) ORDER BY t.id DESC")
-    List<Ticket> findTicketsByUserId(@Param("userId") Integer userId);
+    List<Ticket> findTicketsByUserId(@Param("userId") Long userId);
 }
